Update README.md
Fixing a Program with a Syntax Error

A Syntax Error is an error in the "grammar" of the programming language. These are caught by the compiler and listed out with line number and error found. You will learn how to understand what they tell you with experience. All syntax errors must be corrected before the program will run. If the program runs, this does not mean that it is correct, only that there are no syntax errors. Examples of syntax errors are spelling mistakes in variable names, missing semicolon, unpaired curly braces, etc.

Try running the program. Here we have our first example of the many syntax errors that you no doubt will encounter in this course. The error message you receive may be different depending on the system you are using, but the compiler insists that there is a syntax error somewhere. Unfortunately, where the message indicates that the problem exists, and where the problem actually occurs may be two different places.

To correct the problem place a closing semicolon after the line:

cout << "Today is a great day for Lab"

Most syntax errors are not as easy to spot and correct as this one.

Re-compile the program and when you have no syntax errors, run the program and input 9 when asked (press enter after entering 9). Make sure the answer is correct (18).